+++ src/main/java/monke/Parser.java
@@ -1,8 +1,11 @@
 package monke;
 
 import monke.commands.*;
+import monke.tasks.Deadline;
+import monke.tasks.Event;
+import monke.tasks.Task;
+import monke.tasks.Todo;
 
-import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
