+++ src/main/java/monke/Ui.java
@@ -1,5 +1,7 @@
 package monke;
 
+import monke.tasks.Task;
+
 import java.util.List;
 import java.util.Scanner;
