+++ src/main/java/monke/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package monke;
+package monke.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
