+++ src/main/java/monke/tasks/Event.java
@@ -1,4 +1,4 @@
-package monke;
+package monke.tasks;
 
 public class Event extends Task {
     private String start;
